fix: mobile version
